20.1.2025
Extended the viewModel and Model to include all responses from OpenWeather (fetchWeather).

Added a tracker that tracks the time zone and based on that changes the background of the sheet view.

Extended the WeatherResponse and ViewModel to also include hourly forecasts (fetchForecast), but this does not work as intended yet and has to be investigated.

Changed the "Add City" button to be a toolbar that is injected directly into the CityDetailView instead of being a separate entity in HomePage View. This button is now a variable included in the nested view as needed and displays correctly in the sheet.

Fixed a bug where if the user searched for an empty string, a stack of different city views were displayed in the sheet. Implemented a logic that throws an error popup, notifying the user they should type in a city name.

Added another weather information (clouds) into the CityDetailView.

Found the following issues:
- Cloud info in CityDetailView does not seem to working correctly, it shows 0 on all views, investigation and fix needed.
- Humidity does not seem to work correctly in CityDeatilView either, need to investigate and fix.